refactor with typescript
