+++ src/app.ts
@@ -1,7 +1,7 @@
 import { App } from '@slack/bolt';
 import 'dotenv/config';
 
-const app = new App({
+const app: App = new App({
   token: process.env.SLACK_BOT_TOKEN,
   signingSecret: process.env.SLACK_SIGNING_SECRET,
 });
